Fix bug in crawler about floating numbers on units
It was found that there is a floating number on class units. So,
fields of units in session model was changed.

Also, crawler will print major index in log file and console.
Log analysis will print average and standard deviation of working
times.

+++ lectures/models.py
@@ -48,5 +48,5 @@ class Session(models.Model):
     times = models.ManyToManyField(DayTimeField)
     course = models.ForeignKey(Course, on_delete=models.CASCADE, null=False)
     components = models.ManyToManyField(Category)
-    upper_unit = models.SmallIntegerField(null=False)
-    lower_unit = models.SmallIntegerField(null=False)
+    upper_unit = models.FloatField(null=False)
+    lower_unit = models.FloatField(null=False)
